requirements.txt
numpy==1.26.3
matplotlib==3.8.0
scikit-image==0.22.0
torch==1.13.1
torchvision==0.14.1
Pillow==10.2.0
opencv-python==4.9.0.80
google-cloud-storage==2.14.0
